requirement.txt
flask
tensorflow-cpu == 2.3.0
opencv-python
matplotlib
numpy
pymysql
